+++ removed/trunk/unpackages/apache/apache.spec
@@ -1,33 +1,49 @@
-#New ADVX macros
+%define name	apache
+%define version	1.3.31
+%define release	1sls
+
+# New ADVX macros
 %define ADVXdir %{_datadir}/ADVX
 %{expand:%(cat %{ADVXdir}/ADVX-build)}
 
-%{expand:%%define mm_major %(mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\1/')}
-%{expand:%%define mm_minor %(mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\2/')}
+# the default version of mm required
+%define dmm_major	1
+%define dmm_minor	3.0
+
+%{expand:%%define mm_major %([ -x /usr/bin/mm-config ] && mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\1/' || echo "%{dmm_major}")}
+%{expand:%%define mm_minor %([ -x /usr/bin/mm-config ] && mm-config --version|sed 's/MM \([0-9]\)\.\([0-9.].*\) \(.*\)$/\2/' || echo "%{dmm_minor}")}
 %define mm_version %{mm_major}.%{mm_minor}
 
-%define apache_version 1.3.29
-%define apache_release 1mdk
-%define EAPI_version 2.8.16
-%define modssl_apache_version %{apache_version}
+%define apache_version		%{version}
+%define apache_release		%{release}
+%define EAPI_version		2.8.17
+%define modssl_apache_version	%{apache_version}
 
-Summary:	The most widely used Web server on the Internet.
-Name:		apache
+Summary:	The Apache web server
+Name:		%{name}
 Version:	%{apache_version}
 Release:	%{apache_release}
+License:	Apache License
 Group:		System/Servers
 URL:		http://www.advx.org
-License:	Apache License
-BuildRoot:	%{_tmppath}/%{name}-%{version}-buildroot
-Provides:	webserver 
-Prereq:		apache-modules >= %{apache_version}-%{apache_release}
-Prereq:		apache-conf >= %{apache_version}
-Prereq:		apache-common >= %{apache_version}-%{apache_release}
-Prereq:		rpm-helper
-Prereq:		mm = %{mm_major}.%{mm_minor}
-Provides:	ADVXpackage
-Provides:	AP13package
+Source0:	apache_%{version}.tar.gz
+Source1:	apache_%{version}.tar.gz.asc
+Source2:	http://www.modssl.org/source/mod_ssl-%{EAPI_version}-%{modssl_apache_version}.tar.gz
+Source3:	http://www.modssl.org/source/mod_ssl-%{EAPI_version}-%{modssl_apache_version}.tar.gz.asc
+Source4:	README.ADVX
+Source5:	02_apache.afterboot
+Source6:	apache.run
+Source7:	apache-log.run
+Patch1:		apache_1.3.11-apxs.patch.bz2
+Patch2:		apache_1.3.26-srvroot.patch.bz2
+Patch3:		apache-1.3.23-dbm.patch.bz2
+Patch4:		Configuration.diff.bz2
+Patch5:		apache-1.3.31-baseversion.patch.bz2
+Patch6:		apache-1.3.14-mkstemp.patch.bz2
+Patch8:		apache-1.3.20.manpage.patch.bz2
+Patch9:		apache-1.3.22-man.patch.bz2
 
+BuildRoot:	%{_tmppath}/%{name}-%{version}-buildroot
 BuildConflicts:	BerkeleyDB-devel
 BuildRequires:	ADVX-build >= 1.2
 BuildRequires:	libmm-devel = %{mm_major}.%{mm_minor}
@@ -38,22 +54,14 @@ BuildRequires:	db1-devel
 BuildRequires:	glibc-devel
 BuildRequires:	openssl-devel
 
-Source0:	apache_%{version}.tar.gz
-Source1:	apache_%{version}.tar.gz.asc
-Source2:	http://www.modssl.org/source/mod_ssl-%{EAPI_version}-%{modssl_apache_version}.tar.gz
-Source3:	http://www.modssl.org/source/mod_ssl-%{EAPI_version}-%{modssl_apache_version}.tar.gz.asc
-Source4:	README.ADVX
-
-
-Patch1:		apache_1.3.11-apxs.patch.bz2
-Patch2:		apache_1.3.26-srvroot.patch.bz2
-#Patch3:	apache_1.3.20-nondbm.patch.bz2
-Patch3:		apache-1.3.23-dbm.patch.bz2
-Patch4:		Configuration.diff.bz2
-Patch5:		apache-1.3.29-baseversion.patch.bz2
-Patch6:		apache-1.3.14-mkstemp.patch.bz2
-Patch8:		apache-1.3.20.manpage.patch.bz2
-Patch9:		apache-1.3.22-man.patch.bz2
+PreReq:		apache-modules >= %{apache_version}-%{apache_release}
+PreReq:		apache-conf >= %{apache_version}
+PreReq:		apache-common >= %{apache_version}-%{apache_release}
+PreReq:		rpm-helper afterboot
+PreReq:		mm = %{mm_major}.%{mm_minor}
+Provides:	webserver 
+Provides:	ADVXpackage
+Provides:	AP13package
 
 %description
 Apache is a powerful, full-featured, efficient and freely-available
@@ -69,8 +77,8 @@ format, like PHP4, the Hotwired XSSI module and Apache-ASP. Also included are
 special patches to enable FrontPage 2000 support (see mod_frontpage package).
 
 %package modules
-Group:		System/Servers
 Summary:	Standard modules for Apache
+Group:		System/Servers
 Prereq:		apache-common >= %{apache_version}-%{apache_release}
 Prereq:		mm = %{mm_major}.%{mm_minor}
 Provides:	ADVXpackage
@@ -84,8 +92,8 @@ also with the apache-mod_perl package.
 
 
 %package devel
-Group:		Development/C
 Summary:	Module development tools for Apache 1.3.
+Group:		Development/C
 Provides:	secureweb-devel
 Obsoletes:	secureweb-devel
 Prereq:		mm = %{mm_major}.%{mm_minor}
@@ -110,8 +118,8 @@ you want to be able to compile or develop additional modules
 for it, you'll need to install this package.
 
 %package source
-Group:		System/Servers
 Summary:	Apache Source
+Group:		Development/C
 Prereq:		mm = %{mm_major}.%{mm_minor}
 Prereq:		mm-devel = %{mm_major}.%{mm_minor}
 #No use to install it if you don't have libgdbm.so and libpthread.so!
@@ -129,7 +137,6 @@ The Apache 1.3 Source, including Mandrake patches and EAPI.
 Use this package to build apache-mod_perl, or your own custom version.
 
 %prep
-#unpack apache.
 %setup -q -n apache_%{apache_version}
 %patch1 -p1 
 %patch2 -p1 
@@ -140,6 +147,7 @@ Use this package to build apache-mod_perl, or your own custom version.
 %patch8 -p0
 %patch9 -p0
 
+echo "major: %{mm_major}; minor: %{mm_minor}"
 #Correct layout 
 perl -pi -e 's|" PLATFORM "|%{distribution}/%{apache_release}|;' \
 	src/main/http_main.c
@@ -262,7 +270,7 @@ else
 fi
 
 %install
-rm -rf %{buildroot}
+[ -n "%{buildroot}" -a "%{buildroot}" != / ] && rm -rf %{buildroot}
 mkdir %{buildroot}
 
 # install source
@@ -293,8 +301,16 @@ rm -f %{buildroot}%{_mandir}/man8/logresolve*
 rm -f %{buildroot}%{_mandir}/man8/rotatelogs*
 rm -f %{buildroot}%{_mandir}/man8/apxs*
 
+mkdir -p %{buildroot}%{_datadir}/afterboot
+install -m 0644 %{SOURCE5} %{buildroot}%{_datadir}/afterboot/02_apache
+
+mkdir -p %{buildroot}%{_srvdir}/apache/log
+mkdir -p %{buildroot}%{_srvlogdir}/apache
+install -m 0750 %{SOURCE6} %{buildroot}%{_srvdir}/apache/run
+install -m 0750 %{SOURCE7} %{buildroot}%{_srvdir}/apache/log/run
+
 %clean
-[ "%{buildroot}" != "/" ] && rm -rf %{buildroot} 
+[ -n "%{buildroot}" -a "%{buildroot}" != / ] && rm -rf %{buildroot}
 
 [ "../mod_ssl-%{EAPI_version}-%{modssl_apache_version}" != "/" ] && rm -rf ../mod_ssl-%{EAPI_version}-%{modssl_apache_version}
 
@@ -305,25 +321,32 @@ rm -f %{buildroot}%{_mandir}/man8/apxs*
 %post
 #JMD: do *not* use _post_service here, it is used in apache-conf, since we
 #can have both apache and apache-mod_perl
+%_mkafterboot
 %ADVXpost
 
 %postun
 #JMD: do *not* use _post_service here, otherwise it will uninstall
 #apache-mod_perl as well!!
+%_mkafterboot
 %ADVXpost
 
 %files 
 %defattr(-,root,root)
+%doc README.ADVX
 %{_sbindir}/httpd
 #%{_mandir}/man8/httpd.*
-%doc README.ADVX
+%{_datadir}/afterboot/02_apache
+%dir %{_srvdir}/apache
+%dir %{_srvdir}/apache/log
+%{_srvdir}/apache/run
+%{_srvdir}/apache/log/run
+%dir %attr(0750,nobody,nogroup) %{_srvlogdir}/apache
 
 %files modules
 %defattr(-,root,root)
 %{_libdir}/apache/*.exp
 %{_libdir}/apache/*.ep
 %{_libdir}/apache/*.so
-%doc README.ADVX
 
 %files devel
 %defattr(-,root,root)
@@ -331,14 +354,32 @@ rm -f %{buildroot}%{_mandir}/man8/apxs*
 %{_libdir}/apache/*.a
 %{_sbindir}/apxs
 #%{_mandir}/man8/apxs.*
-%doc README.ADVX
 
 %files source
 %defattr(-,root,root)
 /usr/src/apache_%{apache_version}
-%doc README.ADVX
 
 %changelog
+* Mon May 17 2004 Vincent Danen <vdanen@opensls.org> 1.3.31-1sls
+- 1.3.31
+- mod_ssl 2.8.17
+- rediff P5
+
+* Mon Mar 02 2004 Vincent Danen <vdanen@opensls.org> 1.3.29-4sls
+- supervise scripts
+
+* Wed Feb 11 2004 Vincent Danen <vdanen@opensls.org> 1.3.29-3sls
+- more spec cleaning
+- README.ADVX in only the main package
+- make mm 1.3.0 the default wanted version of mm so we don't get
+  bad output on building and mm isn't installed
+- add afterboot snippet
+- PreReq: afterboot
+
+* Sat Jan 03 2004 Vincent Danen <vdanen@opensls.org> 1.3.29-2sls
+- OpenSLS build
+- tidy spec
+
 * Sat Nov 08 2003 Oden Eriksson <oden.eriksson@kvikkjokk.net> 1.3.29-1mdk
 - apache v1.3.29
 - EAPI v2.8.16
